+++ auth/usecase/user_usecase.go
@@ -6,10 +6,10 @@ import (
 	"log"
 	"time"
 
-	"github.com/metalpoch/olt-blueprint/auth/entity"
 	"github.com/metalpoch/olt-blueprint/auth/model"
 	"github.com/metalpoch/olt-blueprint/auth/repository"
 	"github.com/metalpoch/olt-blueprint/auth/utils"
+	"github.com/metalpoch/olt-blueprint/common/entity"
 	"gorm.io/gorm"
 )
